Split up the sample contract, and fix TH issue
