+++ pluto.cabal
@@ -171,32 +171,39 @@ test-suite spec
 
 
 executable pluto-sample-contract
-    main-is:          Main.hs
+  main-is:          Main.hs
 
-    -- Modules included in this executable, other than Main.
-    -- other-modules:
+  other-modules: 
+    Sample.Offchain
+    Sample.Validator.Haskell
+    Sample.Validator.Pluto
 
-    -- LANGUAGE extensions used by modules in this package.
-    -- other-extensions:
-    build-depends:    
-        base ^>=4.14.3.0,
-        pluto,
-        plutus-contract,
-        plutus-ledger,
-        plutus-ledger-api,
-        plutus-tx,
-        plutus-tx-plugin,
-        lens,
-        text,
-        containers,
-        data-default,
-        freer-extras,
-        hedgehog,
-        tasty,
-        tasty-hedgehog,
-        tasty-quickcheck,
-        QuickCheck
- 
+  -- LANGUAGE extensions used by modules in this package.
+  -- other-extensions:
+  build-depends:    
+      base ^>=4.14.3.0,
+      pluto,
+      plutus-contract,
+      plutus-ledger,
+      plutus-ledger-api,
+      plutus-tx,
+      plutus-tx-plugin,
+      lens,
+      text,
+      containers,
+      data-default,
+      freer-extras,
+      hedgehog,
+      tasty,
+      tasty-hedgehog,
+      tasty-quickcheck,
+      QuickCheck
 
-    hs-source-dirs:   examples/contracts/sample
-    default-language: Haskell2010
+
+  hs-source-dirs:   examples/contracts/sample/src
+  default-language: Haskell2010
+  ghc-options:
+          -- See Plutus Tx readme
+          -Wall -Werror -Wcompat -fwarn-redundant-constraints -fwarn-incomplete-uni-patterns
+          -fwarn-tabs -fwarn-incomplete-record-updates -fwarn-identities -Wmissing-export-lists
+          -fobject-code -fno-ignore-interface-pragmas -fno-omit-interface-pragmas -fno-strictness -fno-ignore-interface-pragmas -fno-specialise
